fix(cs): defer locked chunks in registration sweep
The termination pass of hddRegistrationSweepNext rechecked every
unmarked chunk with hddChunkFindAndLock. MasterConn runs that sweep
from its network event loop, so a chunk held by a worker stopped all
master traffic until the worker released it.

Do not wait there. The termination scan now reports kRetry when an
unmarked Locked chunk remains; MasterConn retains its credit and
retries the sweep one second later from its normal event-loop tick.
It sends the registration end only once no reportable or locked
straggler remains.

RegistrationSweepTests.LockedChunkIsRetriedWithoutWaiting holds a chunk
locked, proves the sweep returns without waiting, then releases it and
checks that the following pass reports it before completing.

Also require mock-disk numeric fields to consume their complete input
ranges, and correct the failover benchmark comment to its actual
10,000,000-chunk default.

+++ src/chunkserver/chunkserver-common/chunk_interface.h
@@ -213,9 +213,9 @@ public:
 	/// Epoch of the last master-driven registration sweep which reported this
 	/// chunk (see hddRegistrationSweep* in hddspacemgr). Guarded by
 	/// gChunksMapMutex, like the registry itself.
-	virtual uint8_t registrationEpoch() const = 0;
+	virtual uint32_t registrationEpoch() const = 0;
 	/// Sets the registration sweep epoch.
-	virtual void setRegistrationEpoch(uint8_t newEpoch) = 0;
+	virtual void setRegistrationEpoch(uint32_t newEpoch) = 0;
 
 	/// Returns the number of blocks in the Chunk.
 	virtual uint16_t blocks() const = 0;
